requirements.txt
python-binance==1.0.19
pyspark==3.3.1
pandas==1.5.3
pyhive
thrift